board:nxp: Set the default Linux partition size to 3G
For OpenIL, it'll mount the whole filesystem (default or Ubuntu)
to the SD card second partition (formatted to Linux style 0x83).
In default the size of the Linux partition is about the size of
the real filesystem. It's too tight for a lot of applications.

So the OpenIL set the default Linux partition to 3G size, this
just effects the SD card partition information, and not increases
the size of final SD card image.

After the booting up of the OpenIL system, there is just 45% space
can be used.

	root@OpenIL-Ubuntu:~# df -h
	Filesystem      Size  Used Avail Use% Mounted on
	/dev/root       208M  106M   88M  55% /

But you can extend the filesystem size to max 3G.

	root@OpenIL-Ubuntu:~# resize2fs /dev/mmcblk0p2
	# or any other number little then 3G
	root@OpenIL-Ubuntu:~# resize2fs /dev/mmcblk0p2 xxx
	# xxx is the size you want set to

After the above command, you can find the filesystem has been extended
to 3G (or other set number).

	root@OpenIL-Ubuntu:~# df -h
	Filesystem      Size  Used Avail Use% Mounted on
	/dev/root       2.9G  106M  2.6G   4% /

This is a balance between the small final image size (considering
program time, OTA time and save the space) and big filesystem
partition for more applications.

Signed-off-by: Liu Gang <Gang.Liu@nxp.com>