board: nxp: ls1021a-tsn: Drop the OP-TEE Secure Boot-enabled build
I don't know of any real users, and the OpenIL documentation on the
topic is just academic material. I don't feel like maintaining more
than a single flavour of OpenIL for a board.

Additionally, having an OS running in the ARM secure world might
interfere with future PREEMPT_RT support in Linux in ways that are
difficult for me to predict at this point.

If you care about this, shout.

Signed-off-by: Vladimir Oltean <olteanv@gmail.com>