board: nxp: ls1021a-tsn: Use community kernel and bootloader
Note: In a way this is a shameless plug, as I am redirecting to very
lightly-patched versions of U-Boot master and Linux net-next which I
myself am maintaining for this board.

But using these versions of the kernel and bootloader is much less of a
maintainership burden than the OpenIL versions, which have forked from
mainline quite a while ago.

Signed-off-by: Vladimir Oltean <olteanv@gmail.com>