board: nxp: ls1021a-tsn: Import U-Boot distro boot env
This is a straight copy-paste from:
https://gitlab.denx.de/u-boot/u-boot/blob/master/include/configs/ls1021atsn.h#L149

See distro boot documentation at:
https://gitlab.denx.de/u-boot/custodians/u-boot-net/blob/master/doc/README.distro

The reason why we're duplicating it by adding it here is that it's
easier to have Buildroot generate an U-Boot environment than to remove
it from the genimage template etc.

But at least remove the ethaddr, eth1addr, eth2addr etc. The MAC
addresses are supposed to be stored using the 'mac' command in the
EEPROM so that they are truly persistent across U-Boot reflashes.

Signed-off-by: Vladimir Oltean <olteanv@gmail.com>