package/linuxptp: Warn if kernel does not support PTP_SYS_OFFSET_EXTENDED
This patch is not submitted upstream, because adoption of this ioctl is
not yet widespread. However, on sja1105 it is critical to get proper
performance with phc2sys, and it is a bit tricky to get recent enough
kernel headers that include this ioctl. So at least warn that something
is wrong.

Signed-off-by: Vladimir Oltean <olteanv@gmail.com>