package/linuxptp: Attempt to deal with frame reordering
It was reported that ptp4l appears to 'freeze' when synchronizing the
sja1105. This turns out to be due to RX reordering of sync/follow-up
frames. This reimplements the ptp4l FSM and adds a new config option:
sync_follow_up_history, which is not yet used in linuxptp.cfg.

Signed-off-by: Vladimir Oltean <olteanv@gmail.com>