board: nxp: ls1021a-tsn: Remove library packages
A lot of junk has accumulated in the board's defconfig - a lot of
packages I don't even know what they are or whether they are needed.

Let's try to only keep top-level (user-visible) packages in the
defconfig, and let those drag other libraries as dependencies. We
shouldn't need to enable the dependencies explicitly.

Signed-off-by: Vladimir Oltean <olteanv@gmail.com>