board: nxp: common: Add script for quick passwordless login
I'm not suggesting you use this, but nonetheless, just in case you want
to shoot yourself in the foot, here's the gun.

Run at your own risk.

For full insecurity, but most practicality, this is best served with a
DHCP reservation (based on MAC address stored in U-Boot EEPROM) and a
rule for ~/.ssh/config such as below:

Host 10.0.0.*
	StrictHostKeyChecking no
	UserKnownHostsFile=/dev/null

Signed-off-by: Vladimir Oltean <olteanv@gmail.com>