board: nxp: ls1028ardb: Add systemd-netword files for net devices
This makes ENETC and the Felix switch ports be brought up at boot time
by systemd. The switch ports are configured to do L2 forwarding by
default.

The double dependency on carrier (swp0-4 -> swp5 -> eno3) is due to the
Felix driver being switchdev in implementation, DSA in reality. The swp5
net device should not exist (it is the feared CPU port from DSA). Until
the Felix driver is fixed, swp5 needs to be a member of br0.

Signed-off-by: Vladimir Oltean <olteanv@gmail.com>