Revert "package/linuxptp: Attempt to deal with frame reordering"
This patch does not really help. When the sja1105 switch driver starts
reordering, it does that due to a bug, and it's completely toast.
The root cause has not been identified yet.

Instead replace that patch with a simpler one, that just notifies the
user that something is extremely wrong. This was silent before.

Signed-off-by: Vladimir Oltean <olteanv@gmail.com>