board: nxp: ls1028a-rdb: Update kernel to November LSDK 19.09 bugfix release
We can now drop the Felix VLAN bugfix patches, as they have been merged
in LSDK already.

There is one new bugfix patch related to Felix MACs at other speeds than
gigabit (10/100). These used to work in BSP 0.3, but there was a
regression during the driver cleanup for LSDK 19.09. The issue was that
at lower speeds than gigabit, the PHY LEDs were flickering like a
Christmas tree because the AN process would not stabilize. The MAC was
also incontrollably starting to send pause frames. It should be left in
gigabit mode in all conditions, since that's the way it was integrated
in the SoC. More details in the patch.

The PCS initialization relies on U-Boot. This works with the current
OpenIL U-Boot (which is based on BSP 0.3), but does not work on LSDK
19.09 (there's another regression there). It _should_ work on the LSDK
19.09 bugfix update, but we're not upgrading to that right now due to
a TFA dependency, which will be solved and integrated in upstream
Buildroot master branch soon [0].

[0]: https://patchwork.ozlabs.org/patch/1198884/

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>