board: nxp: ls1028a-rdb: Use PHY interrupts in poll mode
This is a workaround for some RDB boards (currently suspected to affect
Rev. B, but not known with 100% certainty) where the link status never
comes up on any interface with LSDK 19.09 kernel (neither ENETC nor
Felix).

More details in patch description.

The patch is not applicable for upstream since the Felix switch device
tree was not yet submitted, and ENETC doesn't use interrupts there (it
doesn't even have the MDIO erratum A-050089 mitigation in upstream).

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>