Revert "board: nxp: ls1028a-rdb: Enable the QEMU package"
This reverts commit f2ffb1af0df61124f723bfe68d247339520280d3.

The driver workaround for ENETC MDIO erratum A-050089 requires that the
drivers for all ENETC ports run in the same software partition in order
to efficiently protect against concurrent access to ENETC SI and MDIO
registers. Otherwise there is a real risk that MDIO returns invalid data
(which may manifest as false link loss at high traffic rate, etc).

So virtualization, at least from the perspective of VFIO direct
assignment of certain ports to a VM, would not work very well. And the
QEMU package occupies a lot of space in the rootfs. So remove it.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>