configs/nxp_ls1028ardb-64b: Re-enable PTP timestamping on ENETC
CONFIG_FSL_ENETC_HW_TIMESTAMPING is not enabled in the LSDK 19.09 kernel. When switching OpenIL kernel tags from the previous
OpenIL-linux-201908 tag (based on 4.14) to OpenIL-v1.7-linux-202001 (based on LSDK 19.09 4.19), this config option, previously present in lsdk.config, was lost.

So we need to add a config fragment in Buildroot in order to get the kernel config back for this board.

Fixes: 663bdd1259bd ("OpenIL/LS1028ARDB: TFA support with LSDK1909")
Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>