system/skeleton: add colored prompt and resize command
By default, if we input a line longer than 80 characters in bash, it wraps around
and overwrites the beginning of the line (the prompt) which is annoying
and makes it difficult to keep track of the command we am writing.

The colored prompt is easier for the eyes to follow (it's easier to
distinguish where the output of a long command ends).

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>
Signed-off-by: Jerry Huang <jerry.huang@nxp.com>