package/nxp/qoriq-ptp4l: depend on Linux kernel
As explained in commit 24f79f6ddb05 ("package/nxp/qoriq-ptp4l: use
compiled kernel headers"), linuxptp needs to use the compiled kernel
headers, and for that, the kernel needs to be compiled first. So,
linuxptp should have the kernel as a dependency.

Not sure how come this dependency got missed.

Fixes: 24f79f6ddb05 ("package/nxp/qoriq-ptp4l: use compiled kernel headers")
Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>